Placeing updated code into the Git IN 19-02-2018

+++ src/main/java/com/sermon/mynote/repository/OrganizationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.sermon.mynote.domain.Organization;
 
-public interface OrganizationRepository extends PagingAndSortingRepository<Organization, Long> {
+public interface OrganizationRepository extends PagingAndSortingRepository<Organization, Integer> {
 
 	public List<Organization> findChurchesByUser();
